pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sahitassist"
version = "0.2.0"
description = "SahitAssist: Normalize (Anmol Lipi), segment, and translate Punjabi literary texts with contextual semantic retrieval."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name="Jaap Singh"}
]

dependencies = [
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "tqdm>=4.66",
  "regex>=2024.4.16",
  "rapidfuzz>=3.9",
  "click>=8.1",
  "tenacity>=8.2",
  "backoff>=2.2",
  "httpx>=0.27",
  "orjson>=3.10",
  "python-dotenv>=1.0",
  "rich>=13.7",
  "pdfminer.six>=20240706",
  "typer>=0.12",
  "cachetools>=5.3",
  "filelock>=3.15",
  "numpy>=1.26"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "ruff",
  "mypy",
  "types-PyYAML"
]
vector = [
  "faiss-cpu; platform_system!='Windows'"
]

[project.scripts]
sahitassist = "sahitassist.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = false
warn_unused_configs = true